Changes in template repository manifest
