+++ internal/pkg/template/repository/manifest/manifest.go
@@ -1,20 +1,11 @@
 package manifest
 
 import (
-	"fmt"
-
-	"github.com/keboola/keboola-as-code/internal/pkg/build"
 	"github.com/keboola/keboola-as-code/internal/pkg/filesystem"
-	"github.com/keboola/keboola-as-code/internal/pkg/json"
 	"github.com/keboola/keboola-as-code/internal/pkg/model"
-	"github.com/keboola/keboola-as-code/internal/pkg/utils"
-	"github.com/keboola/keboola-as-code/internal/pkg/validator"
 )
 
-const FileName = `repository.json`
-
 type Manifest struct {
-	fs        filesystem.Fs
 	changed   bool
 	templates []*TemplateRecord
 }
@@ -33,21 +24,10 @@ type VersionRecord struct {
 	model.AbsPath `validate:"dive"`
 }
 
-// file is repository manifest JSON file.
-type file struct {
-	Version   int               `json:"version" validate:"required,min=1,max=2"`
-	Templates []*TemplateRecord `json:"templates"`
-}
-
-func New(fs filesystem.Fs) *Manifest {
+func New() *Manifest {
 	return &Manifest{
-		fs:        fs,
 		templates: make([]*TemplateRecord, 0),
 	}
-}
-
-func Path() string {
-	return filesystem.Join(filesystem.MetadataDir, FileName)
 }
 
 func (m *Manifest) Path() string {
@@ -55,14 +35,14 @@ func (m *Manifest) Path() string {
 }
 
 func Load(fs filesystem.Fs) (*Manifest, error) {
-	// Read manifest file
+	// Load file content
 	manifestContent, err := loadFile(fs)
 	if err != nil {
 		return nil, err
 	}
 
-	// Create manifest struct
-	m := New(fs)
+	// Create manifest
+	m := New()
 	m.templates = manifestContent.Templates
 
 	// Track if manifest was changed after load
@@ -72,81 +52,16 @@ func Load(fs filesystem.Fs) (*Manifest, error) {
 	return m, nil
 }
 
-func (m *Manifest) Save() error {
+func (m *Manifest) Save(fs filesystem.Fs) error {
+	// Create file content
 	content := newFile()
 	content.Templates = m.templates
 
-	// Save manifest file
-	if err := saveFile(m.fs, content); err != nil {
+	// Save file
+	if err := saveFile(fs, content); err != nil {
 		return err
 	}
 
 	m.changed = false
-	return nil
-}
-
-func newFile() *file {
-	return &file{
-		Version:   build.MajorVersion,
-		Templates: make([]*TemplateRecord, 0),
-	}
-}
-
-func loadFile(fs filesystem.Fs) (*file, error) {
-	// Check if file exists
-	path := Path()
-	if !fs.IsFile(path) {
-		return nil, fmt.Errorf("manifest \"%s\" not found", path)
-	}
-
-	// Read JSON file
-	content := newFile()
-	if err := fs.ReadJsonFileTo(path, "manifest", content); err != nil {
-		return nil, err
-	}
-
-	// Fill in parent paths
-	for _, template := range content.Templates {
-		template.AbsPath.SetParentPath(``)
-		for _, version := range template.Versions {
-			version.AbsPath.SetParentPath(template.Path())
-		}
-	}
-
-	// Validate
-	if err := content.validate(); err != nil {
-		return nil, err
-	}
-
-	// Set new version
-	content.Version = build.MajorVersion
-
-	return content, nil
-}
-
-func saveFile(fs filesystem.Fs, manifestContent *file) error {
-	// Validate
-	err := manifestContent.validate()
-	if err != nil {
-		return err
-	}
-
-	// Write JSON file
-	content, err := json.EncodeString(manifestContent, true)
-	if err != nil {
-		return utils.PrefixError(`cannot encode manifest`, err)
-	}
-	file := filesystem.NewFile(Path(), content)
-	if err := fs.WriteFile(file); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func (f *file) validate() error {
-	if err := validator.Validate(f); err != nil {
-		return utils.PrefixError("repository manifest is not valid", err)
-	}
 	return nil
 }
